requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
drf-spectacular>=0.27
psycopg2-binary>=2.9
gunicorn>=21.2
